requirements.txt
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
PyGithub>=2.1.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.8.0
httpx>=0.26.0